requirements.txt
PyJWT
fastapi
uvicorn[standard]
passlib[bcrypt]
bcrypt==4.1.3
pydantic
pydantic-settings
sqlmodel
psycopg2-binary
python-dotenv
emails
alembic